requirements.txt
# Core
torch==2.5.1
torchvision==0.20.1
numpy==2.3.5

# Medical imaging
monai==1.5.2

# Image processing
opencv-python-headless==4.13.0.92
Pillow==12.0.0
albumentations==2.0.8
albucore==0.0.24

# Visualization
matplotlib==3.10.8
seaborn==0.13.2

# Metrics & analysis
scipy==1.17.1
scikit-learn==1.8.0
pandas==3.0.1

# Model analysis
torchinfo==1.8.0
einops==0.8.2

# Experiment tracking
wandb==0.25.1

# Utilities
tqdm==4.67.3
pyyaml==6.0.3

# Optional: Mamba SSM (install separately, requires CUDA)
# pip install mamba-ssm
